+++ package.json
@@ -1,16 +1,17 @@
 {
   "name": "pronote-api",
-  "version": "1.1.0",
-  "description": "Pronote 2020 API",
+  "version": "2.0.0",
+  "description": "Pronote 2020 API - fork from Adrien 'Litarvan' Navratil <adrien1975@live.Fr>",
   "main": "index.js",
   "scripts": {
+    "start": "server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [
     "pronote",
     "api"
   ],
-  "author": "Adrien 'Litarvan' Navratil <adrien1975@live.Fr>",
+  "author": "Ryzzzen",
   "license": "GPL-3.0",
   "dependencies": {
     "axios": "^0.18.1",
